requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.19
mysql-connector-python==8.0.33
Flask-Login==0.6.2
PyJWT==2.8.0
passlib==1.7.4
python-dotenv==1.0.0
stripe==8.5.0
marshmallow==3.19.0
python-dateutil==2.8.2
pandas
matplotlib
seaborn
numpy
PyJWT
passlib
requests